Refactor: move album components to /pages/album. Album header now a separate component.
