+++ src/app/pages/album/Album.tsx
@@ -2,18 +2,18 @@ import { useEffect, useState } from 'react'
 import { useParams } from 'react-router-dom'
 import { useQuery } from '@tanstack/react-query'
 
-import { useAuth } from '../../hooks/useAuth.tsx'
-import { useLibrary } from '../../hooks/useLibrary.tsx'
+import { useAuth } from '../../../hooks/useAuth.tsx'
+import { useLibrary } from '../../../hooks/useLibrary.tsx'
 
-import getAlbum from '../../api/album/getAlbum.ts'
+import getAlbum from '../../../api/album/getAlbum.ts'
 
-import Loading from './Loading'
-import NotFound from './NotFound'
+import Loading from '../Loading'
+import NotFound from '../NotFound'
 
-import AlbumWrapper from '../../components/wrappers/AlbumWrapper.tsx'
-import ActionBar from '../../components/ActionBar.tsx'
-import TrackList from '../../components/songs/TrackList.tsx'
-import Footer from '../../components/Footer.tsx'
+import AlbumWrapper from './components/AlbumWrapper.tsx'
+import ActionBar from '../../../components/ActionBar.tsx'
+import TrackList from '../../../components/songs/TrackList.tsx'
+import Footer from '../../../components/Footer.tsx'
 
 const Album = () => {
     const [colour, setColour] = useState<string | undefined>(undefined) // accepts 'r/g/b' format
